Rearrange code to allow each tab to be its own UserControl.
This allows independent development on individual tabs without muddling up the main form. I reworked Server Details tab to work in this new system. Spell tab has had his code moved into its own control.
