+++ MannikToolbox/Controls/MobControl.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Drawing;
+using System.Data;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace MannikToolbox.Controls
+{
+    public partial class MobControl : UserControl
+    {
+        public MobControl()
+        {
+            InitializeComponent();
+        }
+    }
+}
