+++ MannikToolbox/MannikToolbox.csproj
@@ -78,6 +78,24 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Content\DatabaseManager.cs" />
+    <Compile Include="Controls\MobControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Controls\MobControl.Designer.cs">
+      <DependentUpon>MobControl.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Controls\ServerDetailsControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Controls\ServerDetailsControl.Designer.cs">
+      <DependentUpon>ServerDetailsControl.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Controls\SpellControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Controls\SpellControl.Designer.cs">
+      <DependentUpon>SpellControl.cs</DependentUpon>
+    </Compile>
     <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -93,6 +111,15 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\ConnectionStringService.cs" />
+    <EmbeddedResource Include="Controls\MobControl.resx">
+      <DependentUpon>MobControl.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Controls\ServerDetailsControl.resx">
+      <DependentUpon>ServerDetailsControl.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Controls\SpellControl.resx">
+      <DependentUpon>SpellControl.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
